+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 24.11.2
+
+## Improvements
+
+- [x] using ingress certificate for TLS Ingress Endpoint
+- [x] passing release version during build
+
+## Breaking changes
+
+- [x] remove entrypoint from config, shell will only on the first container of the setup
+
 ## 24.11.1
 
 ## Improvements